==> Chapter_1/02_R-Scripts/01_DataCleaning/11_Agriculture(Globelands).r <==
################################################################################
#### Resampling and Cropping of the Globeland Agriculture Data
################################################################################
# Description: We can use globeland's agricultural fields to augment our
# croplands layer. In this script I therefore extract agricultural fields so we
# can merge them.

# Clear R's brain
rm(list = ls())

# Set the working directory
wd <- "/home/david/Schreibtisch/15. PhD/Chapter_1"
setwd(wd)

# Load required packages
library(raster)   # To handle raster data

# Make use of multicore abilities
beginCluster()

# Import the Globelands dataset
crops <- raster("03_Data/01_RawData/GLOBELAND/Globeland.tif")

# Reclassify the layer so that only cultivated land (Code 10) is kept as 1, the
# rest as 0
rcl <- data.frame(
    old = c(10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 255)
  , new = c(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0))
crops <- reclassify(crops, rcl)

# Load the reference raster
r <- raster("03_Data/02_CleanData/00_General_Raster.tif")

# Aggregate the globelands dataset to match the resolution of the reference
# raster
crops <- aggregate(crops, fact = round(250 / 30), fun = max)

# Resample the layer to match the reference raster
crops_res <- resample(crops, r, "ngb")

# Save the result to file
writeRaster(
    x         = crops_res
  , filename  = "03_Data/02_CleanData/04_AnthropogenicFeatures_Agriculture_Globelands.tif"
  , overwrite = TRUE
)

# End cluster
endCluster()
